Fix error of zero STD for mlc pseudo labels
This commit addresses an issue that arises when with scenes with few frames (or highly noisy scenes), which results in  a standard deviation close to zero.
This issue can also occur when some points are missed at the boundary due to sparse point projection,
resulting in a NaN loss when evaluating the standard deviation.
 To solve this, we applied a blur kernel to the boundary reprojection map to ensure continuity along the wide 360-image.